don't force static runtime for static lib if msvc
It is not "good practice to link to static runtime", quite the opposite actually. It's not bad to link to static runtime, but it's bad practice to force this behavior in a library since consumers should link all their dependencies with the same runtime.
Moreover it defeats principle of least surprise.